Patch from Bart Van Assche <bart.vanassche@gmail.com>:
The patch adds command-line options -m and -q to the script
generate-kernel-patch. These command line options make it possible to
choose at the time the script is invoked whether or not to include the
qla2x00t and/or mpt code in the generated kernel patch.

Please review.

Signed-off-by: Bart Van Assche <bart.vanassche@gmail.com>

Index: scripts/generate-kernel-patch



git-svn-id: svn://svn.code.sf.net/p/scst/svn/trunk@390 d57e44dd-8a1f-0410-8b47-8ef2f437770f