Fix the module parm description so that the name matches the correct name.
Also, this approach changes the default behavior for people who are using
CONFIG_SCST_PROC because the default was that a single host/tgt was added,
but now they have to change their /etc/modules.d/scst.conf or whatever to 
add add_default_tgt=1.

I am not sure that is a good thing.

With the last few commits and this one, I have tested on 2.6.34.1 and 2.6.28
and things seem to work. I am now looking at the problem Vlad reported with
CONFIG_SCST_PROC where local devices show up under SCSI bus or device from 0
to 7.


git-svn-id: svn://svn.code.sf.net/p/scst/svn/trunk@2108 d57e44dd-8a1f-0410-8b47-8ef2f437770f