Add implicit ALUA support. On a setup with more than one target port (sometimes called storage head) this patch makes it possible to tell the initiator which port to use. Since that information can be specified per LUN, with this patch it is not only possible to tell the initiator which path to choose but also to balance the load over target ports. Target port attributes can not only be defined for target ports that exist on the system on which SCST is running but also for target ports present on other nodes of the same HA setup. As far as I know this feature is unique among open source storage target implementations.
Note: with this patch SCSI target port attributes can be defined for each LUN
defined on each SCST target individually. Such an SCST target is either a
physical entity (e.g. a HCA for the ib_srpt driver) or a logical entity (e.g.
an iSCSI target for the iscsi_scst driver). In the last case it is up to the
user to make sure that there is a one-to-one relationship between SCST target
and SCSI target port.

This patch is based on the specifications in the ANSI T10 spc4r30 document.

Signed-off-by: Bart Van Assche <bvanassche@acm.org>



git-svn-id: svn://svn.code.sf.net/p/scst/svn/trunk@3446 d57e44dd-8a1f-0410-8b47-8ef2f437770f